requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
lxml==6.0.2
packaging==26.0
pillow==11.3.0
python-docx==1.2.0
typing_extensions==4.15.0
